Dockerfile/src/requirements.txt
django
tensorflow-cpu
tensorflow-gpu
torch 
torchvision
numpy
sympy
scikit-learn
matplotlib
keras
ipython
pandas
pillow
opencv-python
opencv-contrib-python
nltk
scipy
jupyterlab